# src/main/java/com/skiply/receipt/entity/Receipt.java
package com.skiply.receipt.entity;

	import java.math.BigDecimal;
	import java.time.LocalDateTime;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;


@Entity
@Table(name = "Receipt_transactions")
public class Receipt {
	
	    public Integer getTransactionId() {
		return transactionId;
	}
	public void setTransactionId(Integer transactionId) {
		this.transactionId = transactionId;
	}
	public int getStudentId() {
		return studentId;
	}
	public void setStudentId(int studentId) {
		this.studentId = studentId;
	}
	public BigDecimal getAmount() {
		return amount;
	}
	public void setAmount(BigDecimal amount) {
		this.amount = amount;
	}
	public LocalDateTime getTransactionDate() {
		return transactionDate;
	}
	public void setTransactionDate(LocalDateTime transactionDate) {
		this.transactionDate = transactionDate;
	}
	public String getCardNumber() {
		return cardNumber;
	}
	public void setCardNumber(String cardNumber) {
		this.cardNumber = cardNumber;
	}
	public String getCardType() {
		return cardType;
	}
	public void setCardType(String cardType) {
		this.cardType = cardType;
	}
	public String getReferenceNumber() {
		return referenceNumber;
	}
	public void setReferenceNumber(String referenceNumber) {
		this.referenceNumber = referenceNumber;
	}
	public String getStatus() {
		return status;
	}
	public void setStatus(String status) {
		this.status = status;
	}
		@Id
	    @GeneratedValue(strategy = GenerationType.IDENTITY)
	    private int transactionId;

	    private int studentId;
	    private BigDecimal amount;
	    private LocalDateTime transactionDate;

	    private String cardNumber; // Consider tokenization for security
	    private String cardType;    // e.g., "Visa", "MasterCard"
	    private String referenceNumber; // Unique transaction reference
	    private String status;      // e.g., "Completed", "Pending", "Failed"

	    // Getters and Setters
	}

