# src/main/resources/Fee_API_Spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/receipts/transactions": {
      "post": {
        "tags": [
          "receipt-controller"
        ],
        "summary": "New Fee Receipt",
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Receipt"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/students/{studentId}": {
      "get": {
        "tags": [
          "receipt-controller"
        ],
        "operationId": "getReceipt",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Student_Receipt_Dto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Receipt": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "integer",
            "format": "int32"
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "cardNumber": {
            "type": "string"
          },
          "cardType": {
            "type": "string"
          },
          "referenceNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Student_Receipt_Dto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "studentName": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          },
          "gradeLevel": {
            "type": "integer",
            "format": "int32"
          },
          "transactionId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number"
          },
          "transactionDate": {
            "type": "string"
          },
          "cardType": {
            "type": "string"
          },
          "referenceNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "mobileNumber": {
            "type": "integer",
            "format": "int32"
          },
          "schoolName": {
            "type": "string"
          }
        }
      }
    }
  }
}